Combine context addons; add retry?/retry!
