+++ .gitignore
@@ -7,8 +7,3 @@
 /pkg/
 /spec/reports/
 /tmp/
-
-# Devenv
-.devenv*
-devenv.local.nix
-.direnv*
