--- lib/metaractor/context_has_key.rb
@@ -1,9 +1,0 @@
-module Metaractor
-  module ContextHasKey
-    def has_key?(key)
-      @table.has_key?(key.to_sym)
-    end
-  end
-end
-
-Interactor::Context.send(:include, Metaractor::ContextHasKey)
